Server/src/requirements.txt
fastapi==0.70.1
requests==2.26.0
uvicorn==0.16.0
sqlalchemy==1.3.23
psycopg2==2.9.1
alembic==1.7.7
pyjwt==1.7.1
bcrypt==3.2.0
email-validator==1.1.3
gunicorn==20.1.0
pytest==6.2.5
